pyproject.toml
[project]
name = "modelship"
description = "Ship machine learning models in minutes, not months"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10,<3.14"
dependencies = [
    "jinja2>=3",
    "onnx>=1",
    "onnxruntime>=1",
    "pydantic>=2",
    "pyyaml>=6",
]
keywords = ["machine-learning", "deployment", "webassembly", "onnx"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Other Audience",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development",
    "Topic :: Utilities",
]
dynamic = ["version"]

[project.urls]
Repository = "https://github.com/datalpia/modelship"
Issues = "https://github.com/datalpia/modelship/issues"
Changelog = "https://github.com/datalpia/modelship/blob/main/CHANGELOG.md"

[project.scripts]
modelship = "modelship.__main__:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "bandit==1.8.6",
    "invoke==2.2.1",
    "mypy==1.18.2",
    "pip-audit==2.9.0",
    "pytest==8.4.2",
    "pytest-cov==7.0.0",
    "ruff==0.14.0",
    "types-protobuf==6.32.1.20250918",
    "types-pyyaml==6.0.12.20250915",
]

[tool.hatch.version]
path = "modelship/__about__.py"

[[tool.mypy.overrides]]
module = ["onnxruntime.*"]
ignore_missing_imports = true